ChromiumOS image chromefier by Cristoffer (Linux only for now)

+++ chromefy.sh
@@ -1,33 +1,50 @@
 #!/bin/sh
 
 #Script to install ChromeOS on top of ChromiumOS
-#Must be ran from ChromiumOS Live USB
+#If installing to a system partition, must be ran from ChromiumOS Live USB (otherwise, any Linux distro is fine)
 #You must install ChromiumOS on your Hard Drive before running this script
 
 #Parameters:
-#1 - ChromiumOS (HDD) system partition (ex: /dev/sda3) (I suggest to manually resize it to 4GB before running the script)
+#1 - ChromiumOS image or (HDD) system partition (ex: /dev/sda3) (If partition: I suggest to manually resize it to 4GB before running the script)
 #2 - ChromeOS image file to be installed
 #3 - [Optional] ChromeOS image from older device with TPM1.2 (ex: caroline image file) (needed for TPM2 images like eve to be able to login)
 
 if ( ! test -z {,} ); then echo "Must be ran with \"sudo bash\""; exit 1; fi
 if [ $(whoami) != "root" ]; then echo "Please run with sudo"; exit 1; fi
 if [ -z "$2" ]; then echo "Missing parametes"; exit 1; fi
+if [ -b "$1" ]; then flag_image=false; elif [ -f "$1" ]; then flag_image=true; else echo "Invalid image or partition: $1"; exit 1; fi
+if [ "$flag_image" = false ] && [ ! -e "/usr/sbin/chromeos-install" ]; then echo "You need to be running ChromiumOS (Live USB) to chromefy a drive partition"; exit 1; fi;
 
+if [ ! -d /home/chronos ]; then mkdir /home/chronos; fi
 if [ ! -d /home/chronos/image ]; then mkdir /home/chronos/image; fi
 if [ ! -d /home/chronos/local ]; then mkdir /home/chronos/local; fi
 if [ ! -d /home/chronos/tpm1image ]; then mkdir /home/chronos/tpm1image; fi
+if [ ! -d /home/chronos/RAW ]; then mkdir /home/chronos/RAW; fi
 
-function abort_chromefy {
-    echo "aborting...";
+function cleanup_chromefy {
+    sync
     umount /home/chronos/image 2>/dev/null
     umount /home/chronos/local 2>/dev/null
     umount /home/chronos/tpm1image 2>/dev/null
+    losetup -d "$chromium_image" 2>/dev/null
     losetup -d "$chromeos_image" 2>/dev/null
     losetup -d "$chromeos_tpm1_image" 2>/dev/null
+    rm -rf /home/chronos/RAW/* &>/dev/null
+}
+
+function abort_chromefy {
+    echo "aborting...";
+    cleanup_chromefy
     exit 1
 }
 
 #Checks if images are valid and mounts them
+if [ "$flag_image" = true ]; then
+    chromium_image=`losetup --show -fP "$1"`
+    mount "$chromium_image"p3 /home/chronos/local -o loop,rw  2>/dev/null
+    if [ ! $? -eq 0 ]; then echo "Image $1 does not have a system partition (corrupted?)"; abort_chromefy; fi
+fi
+
 if [ ! -f "$2" ]; then echo "Image $2 not found"; abort_chromefy; fi
 chromeos_image=`losetup --show -fP "$2"`
 mount "$chromeos_image"p3 /home/chronos/image -o loop,ro  2>/dev/null
@@ -40,11 +57,50 @@ if [ ! -z "$3" ]; then
     if [ ! $? -eq 0 ]; then echo "Image $3 does not have a system partition (corrupted?)"; abort_chromefy; fi
 fi
 
+#Increases ROOT_A partition's size (if flashing on image)
+if [ "$flag_image" = true ]; then
+    #Backups ChromiumOS' /lib
+    cp -av /home/chronos/local/lib /home/chronos/RAW/
+    umount /home/chronos/local
+    
+    #Deletes third (ROOT_A) and fifth (ROOT_B) partitions
+    sfdisk --delete $chromium_image 5
+    sfdisk --delete $chromium_image 3
+    
+    #Recreates the fifth partition with 4MB = 4194304
+    START_NEWB=`sgdisk -f $chromium_image`
+    END_NEWB=`expr $START_NEWB + 8192` #Considering sectors of 512 bytes
+    sgdisk -n 5:$START_NEWB:$END_NEWB $chromium_image
+    sgdisk -c 5:"ROOT-B" $chromium_image
+    sgdisk -t 5:"7F01" $chromium_image
+    e2label "$chromium_image"p5 "ROOT-B"
+    mkfs.ext4 "$chromium_image"p5
+    
+    #Recreates the third partition with the remaining space
+    START_NEWA=`sgdisk -f $chromium_image`
+    END_NEWA=`sgdisk -E $chromium_image`
+    sgdisk -n 3:$START_NEWA:$END_NEWA $chromium_image
+    sgdisk -c 3:"ROOT-A" $chromium_image
+    sgdisk -t 3:"7F01" $chromium_image
+    e2label "$chromium_image"p3 "ROOT-A"
+    mkfs.ext4 "$chromium_image"p3
+    
+    #Searches and fixes errors at filesystem-3, then remounts
+    e2fsck -f -y -v -C 0 "$chromium_image"p3
+fi
+
 #Mounts ChromiumOS system partition
-umount "$1" 2>/dev/null
-mkfs.ext4 "$1"
-mount "$1" /home/chronos/local
-if [ ! $? -eq 0 ]; then echo "Partition $1 inexistent"; abort_chromefy; fi
+if [ "$flag_image" = false ]; then
+    chromium_root_dir=""
+    umount "$1" 2>/dev/null
+    mkfs.ext4 "$1"
+    mount "$1" /home/chronos/local
+    if [ ! $? -eq 0 ]; then echo "Partition $1 inexistent"; abort_chromefy; fi
+else
+    chromium_root_dir="/home/chronos/RAW"
+    mount "$chromium_image"p3 /home/chronos/local -o loop,rw,sync  2>/dev/null
+    if [ ! $? -eq 0 ]; then echo "Something went wrong while changing $1 partition table (corrupted?)"; abort_chromefy; fi
+fi
 
 #Copies all ChromeOS system files
 cp -av /home/chronos/image/* /home/chronos/local
@@ -53,8 +109,8 @@ umount /home/chronos/image
 #Copies modules and certificates from ChromiumOS
 rm -rf /home/chronos/local/lib/firmware
 rm -rf /home/chronos/local/lib/modules/ 
-cp -av /lib/firmware /home/chronos/local/lib/
-cp -av /lib/modules/ /home/chronos/local/lib/
+cp -av "$chromium_root_dir"/lib/firmware /home/chronos/local/lib/
+cp -av "$chromium_root_dir"/lib/modules/ /home/chronos/local/lib/
 rm -rf /home/chronos/local/etc/modprobe.d/alsa*.conf
 sed '0,/enforcing/s/enforcing/permissive/' -i /home/chronos/local/etc/selinux/config
 
@@ -92,7 +148,6 @@ if [ ! -z $internal_camera ] && [ ! -z $original_camera ]
     sudo sed -i "s/$original_camera/$internal_camera/" /home/chronos/local/etc/camera/camera_characteristics.conf
 fi
 
-umount /home/chronos/local
-sync
+cleanup_chromefy
 echo
 echo "ChromeOS installed, you can now reboot"
